Allow inclusion of a banner-graphic
The template includes a banner-graphic if a variable named
'BANNER' exists. Otherwise, only the 'SITENAME' is being displayed.

The variable 'BANNER' must be specified in pelicanconf.py
and must contain the filename (including path) of the image,
relative to the site's content/ directory.
E.g: BANNER = 'images/banner.png' if the file is in
content/images/banner.png

It is also ensured via CSS that the banner will be resized to the
site's maximum width, so that too large images don't break the
page.